fix: cleanup pt3
